feature: use chaining build on generate logger builder
