+++ logs.go
@@ -1,5 +1,25 @@
 package welog
 
+import (
+	"fmt"
+	"time"
+)
+
+func (logger *Logger) Log(level LogLevel, message string) {
+	if logger.level >= level {
+		timestamp := time.Now().Format(TimestampFormat)
+		levelString := LevelToString(level, string(logger.mode))
+
+		logMessage := fmt.Sprintf("[%s] [%s] %s\n", timestamp, levelString, message)
+
+		fmt.Fprint(logger.writer, logMessage)
+
+		if logger.logFile != nil && level == Error {
+			fmt.Fprint(logger.logFile, logMessage)
+		}
+	}
+}
+
 func (l *Logger) Error(message string) {
 	l.Log(Error, message)
 }
